+++ frontend/src/pages/Profile/ProfilePosts.tsx
@@ -1,25 +1,21 @@
-import React from 'react';
-import { Grid, Typography } from "@mui/material";
+import React from "react";
+import { Grid } from "@mui/material";
 // @ts-ignore
 import styles from "./Profile.module.css";
 import { PostDto } from "../../types/post/post.dto";
+import { AddPost } from "./AddPost";
+import { Posts } from "./Posts";
 
 interface ProfileInterface {
-    posts?: PostDto[];
+  posts?: PostDto[];
+  userId: string | undefined
 }
 
-export const ProfilePosts: React.FC<ProfileInterface> = ({ posts }) => {
-    return (
-        <Grid sx={{ p: 2 }}>
-            <Typography variant="h5">Posts</Typography>
-            {posts?.map((post) => {
-                return <Grid className={styles.profilePost}>
-                <Typography>{post.title}</Typography>
-                <Typography>{post.content}</Typography>
-                <Typography>Likes count: {post.likesCount}</Typography>
-                    <img src={post.image} />
-                </Grid>
-            })}
-        </Grid>
-    );
+export const ProfilePosts: React.FC<ProfileInterface> = ({ posts, userId }) => {
+  return (
+    <Grid sx={{ p: 2 }}>
+      <AddPost userId={userId}/>
+      <Posts posts={posts}/>
+    </Grid>
+  );
 };