+++ frontend/src/types/post/createPost.dto.ts
@@ -2,5 +2,5 @@ export interface CreatePostDto {
     title: string;
     content: string;
     userId: number;
-    image: string;
+    image?: string;
 }